+++ opentelemetry-sdk/src/opentelemetry/sdk/_logs/_internal/__init__.py
@@ -85,7 +85,7 @@ warnings.simplefilter("once", LogDroppedAttributesWarning)
 
 
 class LogDeprecatedInitWarning(UserWarning):
-    """Custom warning to indicate deprecated LogRecord init was used.
+    """Custom warning to indicate that deprecated and soon to be deprecated Log classes was used.
 
     This class is used to filter and handle these specific warnings separately
     from other warnings, ensuring that they are only shown once without
@@ -104,7 +104,7 @@ class LogLimits:
     This class does not enforce any limits itself. It only provides a way to read limits from env,
     default values and from user provided arguments.
 
-    All limit arguments must be either a non-negative integer, ``None`` or ``LogLimits.UNSET``.
+    All limit arguments must be either a non-negative integer or ``None``.
 
     - All limit arguments are optional.
     - If a limit argument is not set, the class will try to read its value from the corresponding
@@ -125,8 +125,6 @@ class LogLimits:
         max_attribute_length: Maximum length an attribute value can have. Values longer than
             the specified length will be truncated.
     """
-
-    UNSET = -1
 
     def __init__(
         self,
@@ -156,9 +154,6 @@ class LogLimits:
     def _from_env_if_absent(
         cls, value: int | None, env_var: str, default: int | None = None
     ) -> int | None:
-        if value == cls.UNSET:
-            return None
-
         err_msg = "{} must be a non-negative integer but got {}"
 
         # if no value is provided for the limit, try to load it from env
@@ -181,12 +176,6 @@ class LogLimits:
         return value
 
 
-_UnsetLogLimits = LogLimits(
-    max_attributes=LogLimits.UNSET,
-    max_attribute_length=LogLimits.UNSET,
-)
-
-
 class LogRecord(APILogRecord):
     """A LogRecord instance represents an event being logged.
 
@@ -206,7 +195,7 @@ class LogRecord(APILogRecord):
         body: AnyValue | None = None,
         resource: Resource | None = None,
         attributes: _ExtendedAttributes | None = None,
-        limits: LogLimits | None = _UnsetLogLimits,
+        limits: LogLimits | None = None,
         event_name: str | None = None,
     ): ...
 
@@ -226,7 +215,7 @@ class LogRecord(APILogRecord):
         body: AnyValue | None = None,
         resource: Resource | None = None,
         attributes: _ExtendedAttributes | None = None,
-        limits: LogLimits | None = _UnsetLogLimits,
+        limits: LogLimits | None = None,
     ): ...
 
     def __init__(  # pylint:disable=too-many-locals
@@ -242,21 +231,30 @@ class LogRecord(APILogRecord):
         body: AnyValue | None = None,
         resource: Resource | None = None,
         attributes: _ExtendedAttributes | None = None,
-        limits: LogLimits | None = _UnsetLogLimits,
+        limits: LogLimits | None = None,
         event_name: str | None = None,
     ):
-        if trace_id or span_id or trace_flags:
-            warnings.warn(
-                "LogRecord init with `trace_id`, `span_id`, and/or `trace_flags` is deprecated since 1.35.0. Use `context` instead.",
-                LogDeprecatedInitWarning,
-                stacklevel=2,
-            )
-
+        warnings.warn(
+            "LogRecord will be removed in 1.39.0 and replaced by ReadWriteLogRecord and ReadableLogRecord",
+            LogDeprecatedInitWarning,
+            stacklevel=2,
+        )
         if not context:
             context = get_current()
 
+            if trace_id or span_id or trace_flags:
+                warnings.warn(
+                    "LogRecord init with `trace_id`, `span_id`, and/or `trace_flags` is deprecated since 1.35.0. Use `context` instead.",
+                    LogDeprecatedInitWarning,
+                    stacklevel=2,
+                )
+
         span = get_current_span(context)
         span_context = span.get_span_context()
+
+        # Use default LogLimits if none provided
+        if limits is None:
+            limits = LogLimits()
 
         super().__init__(
             **{
@@ -306,7 +304,9 @@ class LogRecord(APILogRecord):
                     dict(self.attributes) if bool(self.attributes) else None
                 ),
                 "dropped_attributes": self.dropped_attributes,
-                "timestamp": ns_to_iso_str(self.timestamp),
+                "timestamp": ns_to_iso_str(self.timestamp)
+                if self.timestamp is not None
+                else None,
                 "observed_timestamp": ns_to_iso_str(self.observed_timestamp),
                 "trace_id": (
                     f"0x{format_trace_id(self.trace_id)}"
@@ -335,6 +335,25 @@ class LogRecord(APILogRecord):
             return attributes.dropped
         return 0
 
+    @classmethod
+    def _from_api_log_record(
+        cls, *, record: APILogRecord, resource: Resource
+    ) -> LogRecord:
+        return cls(
+            timestamp=record.timestamp,
+            observed_timestamp=record.observed_timestamp,
+            context=record.context,
+            trace_id=record.trace_id,
+            span_id=record.span_id,
+            trace_flags=record.trace_flags,
+            severity_text=record.severity_text,
+            severity_number=record.severity_number,
+            body=record.body,
+            attributes=record.attributes,
+            event_name=record.event_name,
+            resource=resource,
+        )
+
 
 class LogData:
     """Readable LogRecord data plus associated InstrumentationLibrary."""
@@ -344,6 +363,11 @@ class LogData:
         log_record: LogRecord,
         instrumentation_scope: InstrumentationScope,
     ):
+        warnings.warn(
+            "LogData will be removed in 1.39.0 and replaced by ReadWriteLogRecord and ReadableLogRecord",
+            LogDeprecatedInitWarning,
+            stacklevel=2,
+        )
         self.log_record = log_record
         self.instrumentation_scope = instrumentation_scope
 
@@ -585,7 +609,7 @@ class LoggingHandler(logging.Handler):
                 )
         return attributes
 
-    def _translate(self, record: logging.LogRecord) -> LogRecord:
+    def _translate(self, record: logging.LogRecord) -> dict:
         timestamp = int(record.created * 1e9)
         observered_timestamp = time_ns()
         attributes = self._get_attributes(record)
@@ -619,17 +643,15 @@ class LoggingHandler(logging.Handler):
             "WARN" if record.levelname == "WARNING" else record.levelname
         )
 
-        logger = get_logger(record.name, logger_provider=self._logger_provider)
-        return LogRecord(
-            timestamp=timestamp,
-            observed_timestamp=observered_timestamp,
-            context=get_current() or None,
-            severity_text=level_name,
-            severity_number=severity_number,
-            body=body,
-            resource=logger.resource,
-            attributes=attributes,
-        )
+        return {
+            "timestamp": timestamp,
+            "observed_timestamp": observered_timestamp,
+            "context": get_current() or None,
+            "severity_text": level_name,
+            "severity_number": severity_number,
+            "body": body,
+            "attributes": attributes,
+        }
 
     def emit(self, record: logging.LogRecord) -> None:
         """
@@ -639,7 +661,7 @@ class LoggingHandler(logging.Handler):
         """
         logger = get_logger(record.name, logger_provider=self._logger_provider)
         if not isinstance(logger, NoOpLogger):
-            logger.emit(self._translate(record))
+            logger.emit(**self._translate(record))
 
     def flush(self) -> None:
         """
@@ -678,11 +700,66 @@ class Logger(APILogger):
     def resource(self):
         return self._resource
 
-    def emit(self, record: LogRecord):
+    @overload
+    def emit(
+        self,
+        *,
+        timestamp: int | None = None,
+        observed_timestamp: int | None = None,
+        context: Context | None = None,
+        severity_number: SeverityNumber | None = None,
+        severity_text: str | None = None,
+        body: AnyValue | None = None,
+        attributes: _ExtendedAttributes | None = None,
+        event_name: str | None = None,
+    ) -> None: ...
+
+    @overload
+    def emit(  # pylint:disable=arguments-differ
+        self,
+        record: APILogRecord,
+    ) -> None: ...
+
+    def emit(
+        self,
+        record: APILogRecord | None = None,
+        *,
+        timestamp: int | None = None,
+        observed_timestamp: int | None = None,
+        context: Context | None = None,
+        severity_text: str | None = None,
+        severity_number: SeverityNumber | None = None,
+        body: AnyValue | None = None,
+        attributes: _ExtendedAttributes | None = None,
+        event_name: str | None = None,
+    ):
         """Emits the :class:`LogData` by associating :class:`LogRecord`
         and instrumentation info.
         """
-        log_data = LogData(record, self._instrumentation_scope)
+
+        # silence deprecation warnings from internal users
+        with warnings.catch_warnings():
+            warnings.simplefilter("ignore", category=LogDeprecatedInitWarning)
+            if not record:
+                record = LogRecord(
+                    timestamp=timestamp,
+                    observed_timestamp=observed_timestamp,
+                    context=context,
+                    severity_text=severity_text,
+                    severity_number=severity_number,
+                    body=body,
+                    attributes=attributes,
+                    event_name=event_name,
+                    resource=self._resource,
+                )
+            elif not isinstance(record, LogRecord):
+                # pylint:disable=protected-access
+                record = LogRecord._from_api_log_record(
+                    record=record, resource=self._resource
+                )
+
+            log_data = LogData(record, self._instrumentation_scope)
+
         self._multi_log_record_processor.on_emit(log_data)