+++ opentelemetry-sdk/tests/test_configurator.py
@@ -72,6 +72,7 @@ from opentelemetry.sdk.trace.sampling import (
     SamplingResult,
     TraceIdRatioBased,
 )
+from opentelemetry.test.mock_test_classes import IterEntryPoint
 from opentelemetry.trace import Link, SpanKind
 from opentelemetry.trace.span import TraceState
 from opentelemetry.util.types import Attributes
@@ -113,7 +114,19 @@ class DummyLogger:
         self.resource = resource
         self.processor = processor
 
-    def emit(self, record):
+    def emit(
+        self,
+        record=None,
+        *,
+        timestamp=None,
+        observed_timestamp=None,
+        context=None,
+        severity_number=None,
+        severity_text=None,
+        body=None,
+        attributes=None,
+        event_name=None,
+    ):
         self.processor.emit(record)
 
 
@@ -293,15 +306,6 @@ class CustomIdGenerator(IdGenerator):
 
     def generate_trace_id(self):
         pass
-
-
-class IterEntryPoint:
-    def __init__(self, name, class_type):
-        self.name = name
-        self.class_type = class_type
-
-    def load(self):
-        return self.class_type
 
 
 class TestTraceInit(TestCase):